optional = goingup && minOccurs != 0
When a BaseObject is extended, in the protobuf/thrift definition
the BaseObject should have all the fields of the StringObject (which extends BaseObject),
but they should be optional.

In the example added "value" is a field declared on the StringObject class
which extends BaseObject.  Both name and the optional description are declared
in the xml definition of BaseObject.

message BaseObject
{
	required string name = 1;
	optional string description = 2;
	optional string value = 3;
}